refactor: extract event components
