+++ src/app/event/[hash]/page.tsx
@@ -1,6 +1,6 @@
 import EventDetailContainer from '../_containers/EventDetailContainer'
 import { getEvent } from '../_services/actions'
-import type { EventType } from '../_containers/EventListContainer'
+import type { EventType } from '../_types'
 
 export default async function EventDetailPage({ params }: { params: { hash: string } }) {
