+++ src/app/event/_types.ts
@@ -1,0 +1,9 @@
+export type EventType = {
+  hash: number
+  link: string
+  title: string
+  content: string | null
+  image?: string
+  html: boolean
+  date: string
+}
